requirements.txt
pandas
numpy
scikit-learn
shap
joblib
streamlit
pyyaml
matplotlib